Add and Improve features + Fix bugs
- Add ewallet for users they can perform basic operations like credit, debit and transfer.

- Add transactions.

- Add Default permission that applies the permission from settings

- Add get_walett() method to User.

- Add CRUD functionality for events

- Add function get_location_from_osm().

- Add list all food for a service endpoint() ---> path('<int:service_pk>/food/',ListRetrieveFoodAPIView.as_view())

- Add reservation functionality "not finished yet"

- Add policy file that will contain the finantial policy. "not finished yet"

+++ EventEase_server/services/views.py
@@ -5,26 +5,30 @@ from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.viewsets import ModelViewSet
 from rest_framework import status
-from rest_framework.exceptions import ValidationError
+from rest_framework.exceptions import ValidationError,PermissionDenied
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.generics import RetrieveUpdateAPIView
-from accounts.permissions import IsAdminUser,IsOwner,IsOwnerOrAdminUser,IsServiceOwnerOrAdmin
+from accounts.permissions import IsAdminUser,IsOwner,IsOwnerOrAdminUser,IsServiceOwnerOrAdmin,Default,DefaultOrIsAdminUser
 # from locations.serializers import LocationSerializer
 from locations.models import Address, Location
 from locations.serializers import AddressSerializer,LocationSerializer
 
-from events.models import EventType
+from events.models import EventType, Event
 
-from .models import (ServiceType,Service,FoodService,ServiceProviderApplication,FavoriteService,FoodTypeService,
-FoodType,FoodServiceFood,Food,DJService,Venue,PhotoGrapherService,EntertainementService,DecorationService,
-Decor,DecorEventType)
+from .models import (ServiceType, Service, FoodService, ServiceProviderApplication, FavoriteService,FoodTypeService,
+FoodType, FoodServiceFood, Food, DJService, Venue, PhotoGrapherService, EntertainementService, DecorationService,
+Decor, DecorEventType, ServiceReservation,Reservation)
 
 from .serializers import (FoodServiceSerializer, ServiceTypeSerializer, ServiceProviderApplicationSerializer,
 FavoriteServiceSerializer, ServiceSerializer, DJServiceSerializer, FoodTypeSerializer, FoodTypeServiceSerializer,
 FoodSerializer, FoodServiceFoodSerializer, VenueSerializer, PhotoGrapherServiceSerializer
 ,EntertainementServiceSerializer, DecorationServiceSerializer, DecorSerializer,DecorEventTypeListSerializer,
-MyServiceTypeSerializer
+MyServiceTypeSerializer, ServiceReservationSerializer
 )
+
+from wallet.models import FEE_PERCENTAGE,CenterWallet
+
+from policy import get_refund_after_cancelling_service_reservation
 
 
 def get_model_for_service_type(type):
@@ -56,6 +60,45 @@ def get_serializer_for_service_type(type):
     if type == 'decoration':
         return DecorationServiceSerializer  
     return None
+
+
+def inside_range(range_start,range_end,start,end):
+    """
+    Check if the line segment [start, end] is fully inside another line segment [range_start, range_end].
+
+    Args:
+        range_start (number): Start of the outer line segment.
+        range_end (number): End of the outer line segment.
+        start (number): Start of the inner line segment.
+        end (number): End of the inner line segment.
+
+    Returns:
+        bool: True if the inner line segment [start, end] is within the outer line segment [range_start, range_end].
+    """
+    if range_start >= range_end or start >= end:
+        return False
+    
+    if start >= range_start and start < range_end and end > range_start and end <= range_end:
+        return True
+    return False
+
+# def inside_event_range(event_start, event_end, reservation_start, resvation_end):
+#     """
+#     Check if the reservation range is inside event range:
+
+#     Args:
+#         event_start (number): Start of the event
+#         event_end (number): End of the event.
+#         reservation_start (number): Start of the reservation.
+#         resvation_end (number): End of the reservation.
+
+#     Returns:
+#         bool: True if the inner line segment [start, end] is within the outer line segment [range_start, range_end].
+#     """
+#     # event_start = event_start.total_seconds()
+#     return inside_range(event_start, event_end, reservation_start, resvation_end)
+
+# print(inside_event_range(datetime(2020,5,2,12,00,00),datetime(2020,5,2,14,00,00),datetime(2020,5,2,12,00,00),datetime(2020,5,2,15,00,00)))
 
 class ServiceTypeViewSet(ModelViewSet):
     queryset = ServiceType.objects.all()
@@ -299,6 +342,8 @@ class MyServiceAPIView(APIView):
     def get(self,requset,**kwargs):
         user = requset.user
         service = Service.objects.filter(service_provider = user).first()
+        model = get_model_for_service_type(service.service_type.type)
+        service = model.objects.filter(service_provider = user).first()
         serializer = get_serializer_for_service_type(service.service_type.type)
         print(serializer)
         serializer = serializer(service)
@@ -333,7 +378,6 @@ class FoodTypeAPIView(APIView):
 
     def get(self,requset,service_pk,**kwargs):
         food_service = FoodService.objects.get(id = service_pk)
-        user = requset.user
         types_of_food = FoodTypeService.objects.filter(foodService = food_service).all()
         print(types_of_food)
         serialzer = FoodTypeServiceSerializer(types_of_food,many = True)
@@ -377,7 +421,7 @@ class FoodAPIView(APIView):
     
     def post(self, request, service_pk, type_pk, **kwargs):
         service = get_object_or_404(FoodService, id=service_pk)
-        food_type = get_object_or_404(FoodType, id=service_pk)
+        food_type = get_object_or_404(FoodType, id=type_pk)
         serializer = FoodSerializer(data=request.data)
         if serializer.is_valid():
             food, created = Food.objects.get_or_create(
@@ -391,13 +435,14 @@ class FoodAPIView(APIView):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def get(self, request, service_pk, food_pk=None, **kwargs):
+    def get(self, request, service_pk,type_pk = None, food_pk=None, **kwargs):
         if food_pk:
             food = get_object_or_404(Food,id = food_pk)
             serializer = FoodSerializer(food)
             return Response(serializer.data, status=status.HTTP_200_OK)            
         service = get_object_or_404(FoodService, id=service_pk)
-        foods = FoodServiceFood.objects.filter(foodService=service)
+        food_type = get_object_or_404(FoodType, id=type_pk)
+        foods = FoodServiceFood.objects.filter(foodService=service,food__food_type=food_type)
         serializer = FoodServiceFoodSerializer(foods, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
     
@@ -413,6 +458,25 @@ class FoodAPIView(APIView):
         elif self.request.method in ['DELETE', 'POST']:
             return [IsServiceOwnerOrAdmin()] 
     
+
+class ListRetrieveFoodAPIView(APIView):
+    def get(self, request, service_pk, food_pk=None, **kwargs):
+
+        food_type_ids = request.query_params.getlist('food_type')
+        service = get_object_or_404(FoodService, id=service_pk)
+        if food_pk:
+            food = get_object_or_404(Food,id = food_pk)
+            try:
+                food = FoodServiceFood.objects.get(foodService=service,food = food)
+            except FoodServiceFood.DoesNotExist:
+                return Response({'message':'this service does not have such food'},status=404)
+            serializer = FoodSerializer(food)
+            return Response(serializer.data, status=status.HTTP_200_OK)         
+        foods = FoodServiceFood.objects.filter(foodService=service)
+        if food_type_ids:
+            foods = foods.filter(food__food_type__in = food_type_ids).distinct()   
+        serializer = FoodServiceFoodSerializer(foods, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
 class LocationDetailView(RetrieveUpdateAPIView):
     serializer_class = LocationSerializer
@@ -479,11 +543,12 @@ class DecorAPIView(APIView):
         decor_event_types_list_serializer = DecorEventTypeListSerializer(data=request.data, partial = True)
         decor_serializer = DecorSerializer(decor, data=request.data, partial=True)
         if decor_event_types_list_serializer.is_valid():
-            DecorEventType.objects.filter(decor = decor).all().delete()
-            decor_event_types = ast.literal_eval(decor_event_types_list_serializer.validated_data.get('decor_event_types')[0])
-            for event_type in decor_event_types:
-                event_type = get_object_or_404(EventType,id = event_type)
-                DecorEventType.objects.get_or_create(decor = decor, event_type = event_type)
+            if decor_event_types_list_serializer.validated_data.get('decor_event_types') is not None:
+                DecorEventType.objects.filter(decor = decor).all().delete()
+                decor_event_types = ast.literal_eval(decor_event_types_list_serializer.validated_data.get('decor_event_types')[0])
+                for event_type in decor_event_types:
+                    event_type = get_object_or_404(EventType,id = event_type)
+                    DecorEventType.objects.get_or_create(decor = decor, event_type = event_type)
 
             if decor_serializer.is_valid():
                 decor_serializer.save()
@@ -503,3 +568,88 @@ class DecorAPIView(APIView):
         elif self.request.method in ['DELETE', 'POST', 'PUT', 'PATCH']:
             return [IsServiceOwnerOrAdmin()] 
     
+
+class ServiceReservationAPIView(APIView):
+    def get(self, request, service_pk, reservation_pk=None):
+        # Retrieve the service
+        service = get_object_or_404(Service, id=service_pk)
+        service_type = service.service_type.type
+        
+        if reservation_pk is not None:
+            # Retrieve a specific reservation
+            reservation = get_object_or_404(ServiceReservation, id=reservation_pk, service=service)
+            serializer = ServiceReservationSerializer(reservation)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        else:
+            reservations = ServiceReservation.objects.filter(service=service)
+            serializer = ServiceReservationSerializer(reservations, many=True)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+        
+
+    def post(self, request, service_pk,reservation_pk=None):
+        service = get_object_or_404(Service,id = service_pk)
+        service_type = service.service_type.type
+        if service_type in ['food', 'decoration']:
+            return Response({"error": "Cannot reserve services of type 'food' or 'decoration'"}, status=status.HTTP_400_BAD_REQUEST)
+        serializer = ServiceReservationSerializer(data = request.data)
+        if serializer.is_valid():
+            start_time = serializer.validated_data.get('start_time')
+            end_time = serializer.validated_data.get('end_time')
+            event = serializer.validated_data.get('event')
+            if event.user != request.user:
+                return Response({'message':'you do not own this event'}, status = status.HTTP_400_BAD_REQUEST)
+            # check if it is in the event time range
+            is_inside = inside_range(event.start_time,event.end_time,start_time,end_time)
+            if not is_inside:
+                return Response({'message':'make sure that start time is smaller than end time or the reservation range is inside the event range'},status=status.HTTP_400_BAD_REQUEST)
+            model = get_model_for_service_type(service_type)
+            service = model.objects.get(id = service_pk)
+            print(service.hourly_rate)
+            duration = (end_time - start_time).total_seconds() / 3600
+            cost = float(service.hourly_rate) * duration 
+            if cost + cost * FEE_PERCENTAGE >= request.user.get_wallet().balance:
+                return Response({'messgae':'insuficient funds'},status=status.HTTP_400_BAD_REQUEST)
+            service_reservation = serializer.save(service = service,cost = cost)
+            request.user.get_wallet().transfer_funds(target_wallet = CenterWallet.objects.first(),amount = cost, reservation = service_reservation)
+
+
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+        else:
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        
+
+    # def delete(self, request, service_pk,reservation_pk=None):
+    #     reservation = get_object_or_404(ServiceReservation, id = reservation_pk)
+    #     # reservation_owner = reservation.event.user
+    #     print(request.user.is_superuser)
+    #     if reservation.event.user == request.user or request.user.is_superuser:
+    #         reservation.delete()
+    #         return Response(status=status.HTTP_204_NO_CONTENT)
+    #     print('m')
+    #     return Response({'message':'you do not have permission to perform this action'}, status=status.HTTP_403_FORBIDDEN)
+
+    def get_permissions(self):
+        if self.request.method in ['GET', 'POST']:
+            return [DefaultOrIsAdminUser()]
+        elif self.request.method == 'DELETE':
+            return [DefaultOrIsAdminUser()] 
+        
+
+class RejectServiceReservationAPIView(APIView):
+    def post(self, request,service_pk=None, reservation_pk=None):
+        reservation = get_object_or_404(ServiceReservation, id = reservation_pk)
+        if request.user == reservation.service.service_provider:
+            reservation.status = 'Rejected'
+            reservation_cost = float(reservation.cost) 
+            fee = reservation_cost* FEE_PERCENTAGE
+            CenterWallet.objects.first().transfer_funds( reservation.event.user.get_wallet(),amount=reservation_cost+fee,reservation=reservation)   
+            reservation.save()
+            return Response({'message':'the reservation has been rejected successfully'}, status= 200)  
+        else:
+            raise PermissionDenied('you do not have permission to perform this action')
+
+
+class ConfirmReservaiton(APIView):
+    # when service provider confirm 
+    # service_provider.get_wallet().transfer_funds(target_wallet = CenterWallet.objects.first(),amount = cost * 0.5, reservation = service_reservation)
+    pass
